requirements.txt
openai
gradio
requests
PyPDF2
python-dotenv